[#545] Extract scrub functions to shared module + test production code
Address RE1 review: extract scrubSecrets/scrubScrollback into
server/scrub-secrets.js so tests import the production code path
instead of duplicating the functions. Add 3 integration-path tests
that verify server/index.js imports from and calls the shared module.

28/28 tests pass.
